Complete hero section structure and styles
Initial completion for a structure and styling closely following and
adhering to design specifications. Used a fair amount of time to
match elements as per instructions and Googled some edge cases such
as using a wrapper container instead of margins to line-up hero section
closely with navigation.

Modified README.md for some typos and did light corrections to goals.